Extract behave interface
