+++ biobox_cli/command/verify.py
@@ -13,16 +13,8 @@ Available Biobox types:
   short_read_assembler  Assemble short reads into contigs
 """
 
-import biobox_cli.util as util
-import sys, os, os.path, tempfile, json
-
-def verification_file(biobox):
-    from pkg_resources import resource_filename
-    file_ = biobox + '.feature'
-    return os.path.abspath(resource_filename(__name__, os.path.join('..', 'verification', file_)))
-
-def tmp_feature_dir():
-    return os.path.abspath(os.path.join(os.getcwd(), 'biobox_verify'))
+import biobox_cli.util             as util
+import biobox_cli.behave_interface as behave
 
 def run(argv):
     opts   = util.parse_docopt(__doc__, argv, False)
@@ -30,20 +22,7 @@ def run(argv):
     image  = opts['<image>']
     task   = opts['--task']
 
-    from behave.__main__ import main as behave_main
-    _, tmp_file = tempfile.mkstemp()
-    cmd = "{file} --define IMAGE={image} --define TASK={task} --define TMPDIR={tmp_dir} --outfile {tmp_file} --format json.pretty --no-summary --stop"
-    args = {'file':     verification_file(biobox),
-            'tmp_dir':  tmp_feature_dir(),
-            'image':    image,
-            'tmp_file': tmp_file,
-            'task':     task}
+    results = behave.run(biobox, image, task)
 
-    behave_main(cmd.format(**args))
-
-    with open(tmp_file, 'r') as f:
-        features = json.loads(f.read())
-
-
-    if "failed" in map(lambda i: i['status'], features):
+    if "failed" in map(lambda i: i['status'], results):
         util.err_exit('failed_verification', {'image': image, 'biobox': biobox.replace('_', ' ')})
